requirements.txt
Pillow
imageio
networkx
numpy
opencv-python
graphviz
scikit-image
scipy
svgwrite
webcolors
matplotlib
torch
torchvision
Werkzeug==0.16.0
Flask
flask_uploads
pickle-mixin
flask-cors